Restructure of the GUI to utilize threading
This also makes the code a LOT more stable. The PID controller calculations
within an ISR were causing the microcontroller to periodicially crash, as the ISRs
were taking too long to complete and were backing up. Instead, a flag is now set
so that the background processing can occur without blocking interrupts.